+++ packages/util/src/hex/index.ts
@@ -9,6 +9,7 @@ export { hexAddPrefix } from './addPrefix';
 export { hexFixLength } from './fixLength';
 export { hexHasPrefix } from './hasPrefix';
 export { hexStripPrefix } from './stripPrefix';
+export { hexToBigInt } from './toBigInt';
 export { hexToBn } from './toBn';
 export { hexToNumber } from './toNumber';
 export { hexToString } from './toString';