+++ packages/util/src/hex/stripPrefix.spec.ts
@@ -22,6 +22,12 @@ describe('hexStripPrefix', (): void => {
     ).toEqual('1223');
   });
 
+  it('strips the prefix from hex strings (non 2 lnegth)', (): void => {
+    expect(
+      hexStripPrefix('0x123')
+    ).toEqual('123');
+  });
+
   it('returns un-prefixed hex as-is', (): void => {
     expect(
       hexStripPrefix('abcd1223')