+++ packages/util/src/is/buffer.ts
@@ -1,6 +1,8 @@
 // Copyright 2017-2021 @polkadot/util authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { hasBuffer } from '../has';
+
 /**
  * @name isBuffer
  * @summary Tests for a `Buffer` object instance.
@@ -16,5 +18,5 @@
  * ```
  */
 export function isBuffer (value: unknown): value is Buffer {
-  return typeof Buffer !== 'undefined' && Buffer.isBuffer(value);
+  return hasBuffer && Buffer.isBuffer(value);
 }