+++ packages/util-crypto/src/base32/bs32.ts
@@ -1,6 +1,16 @@
 // Copyright 2017-2021 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-// https://github.com/multiformats/js-multibase/blob/master/src/constants.js
+import { utils } from 'micro-base';
+
 export const BASE32_ALPHABET = 'abcdefghijklmnopqrstuvwxyz234567';
-export const BITS_PER_CHAR = 5;
+
+// We define our own chain, the default base32 has padding
+export const base32 = utils.chain(
+  utils.radix2(5),
+  utils.alphabet(BASE32_ALPHABET),
+  {
+    decode: (input: string) => input.split(''),
+    encode: (input: string[]) => input.join('')
+  }
+);