+++ packages/util/src/assert.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { isFunction } from './is/function';
+import { isNull } from './is/null';
 import { isUndefined } from './is/undefined';
 
 type MessageFn = () => string;
@@ -36,8 +37,8 @@ export function assert (condition: unknown, message: string | MessageFn): assert
  * @name assertReturn
  * @summart Returns when the value is not undefined, otherwise throws assertion error
  */
-export function assertReturn <T> (value: T | undefined, message: string | MessageFn): T {
-  assert(!isUndefined(value), message);
+export function assertReturn <T> (value: T | undefined | null, message: string | MessageFn): T {
+  assert(!isUndefined(value) && !isNull(value), message);
 
   return value;
 }