+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes:
 
 - Add `stringPascalCase` utility
 - Add strongly typed `object{Entries, Keys, Values}` utilities
+- Adjust util-crypto BN conversion to use named constants
 - Expand tests for uncovered functions