+++ CHANGELOG.md
@@ -4,14 +4,16 @@
 
 Changes:
 
-- Add `stringPascalCase` utility
 - Add strongly typed `object{Entries, Keys, Values}` utilities
 - Add `{hex, u8a}ToBigInt` utilities
-- Expand tests for uncovered functions
-- Sync with upstream Substrate ss58 registry
+- Add `bi*` utilities for `bigint` operations (mirror of `bn*`)
+- Add `stringPascalCase` utility
+- Add `isCodec` to check for Codec-like structures
 - Optimize for `hexToU8a` and `u8aToHex` functions
 - Optimize `u8aEq` to use `DataView` for compares
 - Space optimization for bip39 wordlists
+- Sync with upstream Substrate ss58 registry
+- Expand tests for uncovered functions
 
 
 ## 7.8.2 Nov 7, 2021