+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 - Add `bi*` utilities for `bigint` operations (mirror of `bn*`)
 - Add `stringPascalCase` utility
 - Add `isCodec` to check for Codec-like structures
+- Add `has*` detection shortcuts (e.g. `hasWasm` or `hasBuffer`)
 - Optimize for `hexToU8a` and `u8aToHex` functions
 - Optimize `u8aEq` to use `DataView` for compares
 - Space optimization for bip39 wordlists