+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changes:
 - Add `stringPascalCase` utility
 - Add `isCodec` to check for Codec-like structures
 - Add `has*` detection shortcuts (e.g. `hasWasm` or `hasBuffer`)
+- Add `sha256AsU8a` util (& replace use internally)
 - Optimize for `hexToU8a` and `u8aToHex` functions
 - Optimize `u8aEq` to use `DataView` for compares
 - Align base{32, 58, 64} around (audited) `micro-base`